Configurando pagina por defecto y agregando alertas sweetalert2

+++ src/app/components/cart-app.component.ts
@@ -6,6 +6,7 @@ import { CartItem } from '../models/cartItem';
 import { NavbarComponent } from './navbar/navbar.component';
 import { Router, RouterOutlet } from '@angular/router';
 import { SharingDataService } from '../services/sharing-data.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'cart-app',
@@ -57,24 +58,47 @@ export class CartAppComponent implements OnInit{
       this.router.navigate(['/cart'], {
       state: {items: this.items, total: this.total}
       })
+
+      Swal.fire({
+        title: "Shopping",
+        text: "Producto agregado al carrito",
+        icon: "success"
+      });
     })
+
   }
 
   onDeleteCart(): void{
     this.sharingDataService.idProductEventEmitter.subscribe(id =>{
-      this.items = this.items.filter(item => item.product.id !== id);
-      if(this.items.length == 0){
-        sessionStorage.removeItem('cart');
-        sessionStorage.clear();
-      }
-      this.calculateTotal();
-      this.saveSession();
-
-      this.router.navigateByUrl('/', {skipLocationChange: true}).then(() => {
-        this.router.navigate(['/cart'], {
-          state: {items: this.items, total: this.total}
-        })
-      })
+      Swal.fire({
+        title: "¿Estas seguro que quieres eliminar este producto del carrito?",
+        text: "No podrás deshacer este cambio",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Si, eliminar."
+      }).then((result) => {
+        if (result.isConfirmed) {
+          this.items = this.items.filter(item => item.product.id !== id);
+          if(this.items.length == 0){
+            sessionStorage.removeItem('cart');
+            sessionStorage.clear();
+          }
+          this.calculateTotal();
+          this.saveSession();    
+          this.router.navigateByUrl('/', {skipLocationChange: true}).then(() => {
+            this.router.navigate(['/cart'], {
+              state: {items: this.items, total: this.total}
+            })
+          })
+          Swal.fire({
+            title: "Eliminado!",
+            text: "El producto ha sido eliminado.",
+            icon: "success"
+          });
+        }
+      });
     })
   }
